Resolve conflicts

+++ src/attackplane.hpp
@@ -6,7 +6,7 @@ class AttackPlane: public Enemy {
 private:
 
 public:
-    AttackPlane(double speed, int value, int price, int hp, sf::Sprite sprite, Game& game, Pos position)
+    AttackPlane(int speed, int value, int price, int hp, sf::Sprite sprite, Game& game, Pos position)
             : Enemy(speed, value, price, hp, sprite, game, position) {};
 
     void update();
