+++ src/enemy.hpp
@@ -9,7 +9,7 @@ protected:
     int price_;
     int maxHP_;
 public:
-    explicit Enemy(double speed, int value, int price, int maxHP, sf::Sprite sprite, Game& game, Pos position)
+    explicit Enemy(int speed, int value, int price, int maxHP, sf::Sprite sprite, Game& game, Pos position)
             : GameObject(sprite, game, position), speed_(speed), value_(value), price_(price), maxHP_(maxHP) {};
     
     virtual ~Enemy() {};
