+++ src/tank.hpp
@@ -6,7 +6,7 @@ class Tank: public Enemy {
 private:
 
 public:
-    Tank(double speed, int value, int price, int hp, sf::Sprite sprite, Game& game, Pos position)
+    Tank(int speed, int value, int price, int hp, sf::Sprite sprite, Game& game, Pos position)
         : Enemy(speed, value, price, hp, sprite, game, position) {};
 
     void update();
